16-LB2
The new version is able to create users and delete users, here I used the sqllights which is really nice to work with. Now you can use SQL to show your users as you want and you even have a DB which is great. Not only that all the data is being held in DB so nothing is secured in the application itself.